# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  student: "student",
} as const;

export interface AuthUser {
  id: number;
  username: string;
  fullName: string;
  role: AuthUserRole;
  active: boolean;
  neverExpires: boolean;
  quizOnce: boolean;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export type AdminUser = AuthUser & {
  password: string;
  createdAt: string;
};

export type UserInputRole = (typeof UserInputRole)[keyof typeof UserInputRole];

export const UserInputRole = {
  admin: "admin",
  student: "student",
} as const;

export interface UserInput {
  username: string;
  password: string;
  fullName: string;
  role: UserInputRole;
  active: boolean;
  neverExpires: boolean;
  quizOnce: boolean;
}

export interface DashboardStats {
  attemptsCount: number;
  bestScore: number;
  lastScore: number | null;
  canTakeQuiz: boolean;
  lockReason: string | null;
}

export interface Question {
  id: number;
  question: string;
  options: string[];
  correctAnswer: number;
  explanation: string;
  imageQuestion: string | null;
}

export interface SubmitAnswer {
  questionId: number;
  selectedAnswer: number;
}

export interface SubmitAttemptRequest {
  answers: SubmitAnswer[];
}

export interface AttemptResult {
  id: number;
  score: number;
  total: number;
  percentage: number;
  passed: boolean;
  createdAt: string;
}

export interface ScoreboardEntry {
  rank: number;
  username: string;
  fullName: string;
  bestScore: number;
  attemptsCount: number;
  lastScore: number | null;
}

export interface AdminResult {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  score: number;
  total: number;
  percentage: number;
  passed: boolean;
  createdAt: string;
}