# lib/api-zod/src/generated/types/adminResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminResult {
  id: number;
  userId: number;
  username: string;
  fullName: string;
  score: number;
  total: number;
  percentage: number;
  passed: boolean;
  createdAt: string;
}