# lib/api-zod/src/generated/types/attemptResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptResult {
  id: number;
  score: number;
  total: number;
  percentage: number;
  passed: boolean;
  createdAt: string;
}