# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from "./authUserRole";

export interface AuthUser {
  id: number;
  username: string;
  fullName: string;
  role: AuthUserRole;
  active: boolean;
  neverExpires: boolean;
  quizOnce: boolean;
}