# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  attemptsCount: number;
  bestScore: number;
  lastScore: number | null;
  canTakeQuiz: boolean;
  lockReason: string | null;
}