# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserInputRole } from "./userInputRole";

export interface UserInput {
  username: string;
  password: string;
  fullName: string;
  role: UserInputRole;
  active: boolean;
  neverExpires: boolean;
  quizOnce: boolean;
}